[IMP] business_product_location : update test readability / use folder report / update __openerp__.py
